@@ -5,7 +5,7 @@
         }
     },
     "require": {
-        "php": "^7.3",
+        "php": "^7.4",
         "ext-pdo": "*",
         "ext-json": "*"
     },
