@@ -9,7 +9,7 @@ class ConnectionCreator
     public static function getConnection(): \PDO
     {
         if (is_null(self::$pdo)) {
-            $caminhoBanco = __DIR__ . '/../../banco.sqlite';
+            $caminhoBanco = __DIR__ . '../../banco.sqlite';
             self::$pdo = new \PDO('sqlite:' . $caminhoBanco);
             self::$pdo->setAttribute(\PDO::ATTR_ERRMODE, \PDO::ERRMODE_EXCEPTION);
         }
